requirements.txt
djangorestframework
pymysql
reportlab
django==4.2
pillow
openpyxl